refactor(visual-editing): move patch resolvers to util file
