+++ packages/visual-editing/src/ui/context-menu/contextMenuItems.ts
@@ -16,23 +16,19 @@ import {
   SortIcon,
   UnpublishIcon,
 } from '@sanity/icons'
-import {at, insert, truncate, type NodePatchList} from '@sanity/mutate'
-import {get} from '@sanity/util/paths'
 import type {ContextMenuNode, OverlayElementField, OverlayElementParent} from '../../types'
 import {getNodeIcon} from '../../util/getNodeIcon'
-import {getArrayItemKeyAndParentPath} from '../../util/mutations'
-import {randomKey} from '../../util/randomKey'
+import {
+  getArrayDuplicatePatches,
+  getArrayInsertPatches,
+  getArrayMovePatches,
+  getArrayRemovePatches,
+} from '../../util/mutations'
 import type {OptimisticDocument} from '../optimistic-state/useDocuments'
 
 export function getArrayRemoveAction(node: SanityNode, doc: OptimisticDocument): () => void {
   if (!node.type) throw new Error('Node type is missing')
-  return () =>
-    doc.patch(({snapshot}) => {
-      const {path: arrayPath, key: itemKey} = getArrayItemKeyAndParentPath(node)
-      const array = get(snapshot, arrayPath) as {_key: string}[]
-      const currentIndex = array.findIndex((item) => item._key === itemKey)
-      return [at(arrayPath, truncate(currentIndex, currentIndex + 1))]
-    })
+  return () => doc.patch(({snapshot}) => getArrayRemovePatches(node, snapshot))
 }
 
 function getArrayInsertAction(
@@ -42,70 +38,12 @@ function getArrayInsertAction(
   position: 'before' | 'after',
 ): () => void {
   if (!node.type) throw new Error('Node type is missing')
-  return () =>
-    doc.patch(() => {
-      const {path: arrayPath, key: itemKey} = getArrayItemKeyAndParentPath(node)
-      const insertKey = randomKey()
-      const referenceItem = {_key: itemKey}
-      return [
-        at(arrayPath, insert([{_type: insertType, _key: insertKey}], position, referenceItem)),
-      ]
-    })
-}
-
-function getArrayMovePatches(
-  node: SanityNode,
-  doc: OptimisticDocument,
-  moveTo: 'previous' | 'next' | 'first' | 'last',
-): NodePatchList {
-  if (!node.type) throw new Error('Node type is missing')
-  const {path: arrayPath, key: itemKey} = getArrayItemKeyAndParentPath(node)
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore - Type instantiation is excessively deep and possibly infinite.
-  const array = doc.get(arrayPath) as {_key: string}[]
-  const item = doc.get(node.path)
-  const currentIndex = array.findIndex((item) => item._key === itemKey)
-
-  let nextIndex = -1
-  let position: 'before' | 'after' = 'before'
-
-  if (moveTo === 'first') {
-    if (currentIndex === 0) return []
-    nextIndex = 0
-    position = 'before'
-  } else if (moveTo === 'last') {
-    if (currentIndex === array.length - 1) return []
-    nextIndex = -1
-    position = 'after'
-  } else if (moveTo === 'next') {
-    if (currentIndex === array.length - 1) return []
-    nextIndex = currentIndex
-    position = 'after'
-  } else if (moveTo === 'previous') {
-    if (currentIndex === 0) return []
-    nextIndex = currentIndex - 1
-    position = 'before'
-  }
-
-  return [
-    at(arrayPath, truncate(currentIndex, currentIndex + 1)),
-    at(arrayPath, insert(item, position, nextIndex)),
-  ]
+  return () => doc.patch(() => getArrayInsertPatches(node, insertType, position))
 }
 
 function getDuplicateAction(node: SanityNode, doc: OptimisticDocument): () => void {
   if (!node.type) throw new Error('Node type is missing')
-
-  return () =>
-    doc.patch(({snapshot}) => {
-      const {path: arrayPath, key: itemKey} = getArrayItemKeyAndParentPath(node)
-
-      const item = get(snapshot, node.path) as object
-      const duplicate = {...item, _key: randomKey()}
-
-      return [at(arrayPath, insert(duplicate, 'after', {_key: itemKey}))]
-    })
+  return () => doc.patch(({snapshot}) => getArrayDuplicatePatches(node, snapshot))
 }
 
 export function getContextMenuItems(context: {
